Задача 2 сделана.  Вывести на консоль все методы класса, включая все родительские методы (включая приватные).
